[project-test] Move to E2E test to avoid interference with existing tests
